requirements.txt
anthropic
mss
SpeechRecognition
Pillow
rich
python-dotenv
